+++ pymoose/pymoose/predictors/__init__.py
@@ -1,6 +1,6 @@
 from pymoose.predictors.linear_predictor import LinearClassifier
 from pymoose.predictors.linear_predictor import LinearRegressor
-from pymoose.predictors.onnx import from_onnx
+from pymoose.predictors.transcribe_onnx import from_onnx
 from pymoose.predictors.tree_ensemble import TreeEnsembleClassifier
 from pymoose.predictors.tree_ensemble import TreeEnsembleRegressor
