+++ pymoose/pymoose/predictors/transcribe_onnx_test.py
@@ -4,12 +4,13 @@ import pathlib
 
 from absl.testing import absltest
 from absl.testing import parameterized
-from onnx import load_model
+import onnx as ox
 
 from pymoose.logger import get_logger
 from pymoose.predictors import linear_predictor
-from pymoose.predictors import onnx
+from pymoose.predictors import transcribe_onnx
 from pymoose.predictors import tree_ensemble
+from pymoose import edsl
 
 _SK_MODELS = [
     ("linear_regression", linear_predictor.LinearRegressor),
@@ -26,13 +27,13 @@ class PredictoOnnxTest(parameterized.TestCase):
         root_path = pathlib.Path(__file__).parent.absolute()
         fixture_path = root_path / "fixtures" / f"{onnx_fixture}.onnx"
         with open(fixture_path, "rb") as model_fixture:
-            onnx_proto = load_model(model_fixture)
+            onnx_proto = ox.load_model(model_fixture)
         return onnx_proto
 
     @parameterized.parameters(*_SK_MODELS)
     def test_regression_logic(self, model_name, predictor_cls):
         model_proto = self._load_onnx(model_name)
-        model = onnx.from_onnx(model_proto)
+        model = transcribe_onnx.from_onnx(model_proto)
         assert isinstance(model, predictor_cls)
